Rename claimantsStore to claimant
