+++ src/services/stellar/operations/createClaimableBalanceOperation.ts
@@ -1,6 +1,6 @@
 import { get } from 'svelte/store';
-import { claimantsStore } from '../../../utils/stores/claimantsStore';
+import { claimant } from '../../../utils/stores/claimant';
 
 export function buildCreateClaimableBalanceOperation(formData: FormData): any {
-  console.log(get(claimantsStore));
+  console.log(get(claimant));
 }
